Make the tests for Monte Carlo pass

+++ src/com/sillysoft/lux/agent/utils/simulation/SimulationResult.java
@@ -5,12 +5,12 @@ public class SimulationResult {
     public final boolean victory;
 
     // The average number of attacker armies, averaged from every simulation
-    public final int averageArmiesRemaining;
+    public final double averageArmiesRemaining;
 
     // The percent of simulations that resulted in a win
     public final double percentChanceAnyWin;
 
-    public SimulationResult(boolean victory, int averageArmiesRemaining, double percentChanceAnyWin) {
+    public SimulationResult(boolean victory, double averageArmiesRemaining, double percentChanceAnyWin) {
         this.victory = victory;
         this.averageArmiesRemaining = averageArmiesRemaining;
         this.percentChanceAnyWin = percentChanceAnyWin;
